Added a requires decorator to the Metadata class.
The requires decorator makes metadata restrictions more explicit. If a 
certain function requires a certain metadata field(s) to present, just 
add a requires decorator to it, with the name of that field(s) in a 
list, and voila! Also used it for the first time in the tseries module.